+++ src/main/java/dev/vality/testcontainers/annotations/kafka/EmbeddedKafkaTestExtension.java
@@ -39,7 +39,7 @@ public class EmbeddedKafkaTestExtension implements BeforeEachCallback {
 
         var applicationContext = SpringExtension.getApplicationContext(context);
         var embeddedKafkaBroker = applicationContext.getBean(EmbeddedKafkaBroker.class);
-        try (var adminClient = AdminClient.create(adminProperties(embeddedKafkaBroker))) {
+        try (var adminClient = AdminClient.create(buildAdminProperties(embeddedKafkaBroker))) {
             var existingTopics = adminClient.listTopics().names().get(WAIT_TIMEOUT_SECONDS, TimeUnit.SECONDS);
             var missingTopics = topics.stream()
                     .filter(topic -> !existingTopics.contains(topic))
@@ -47,19 +47,19 @@ public class EmbeddedKafkaTestExtension implements BeforeEachCallback {
             if (!missingTopics.isEmpty()) {
                 embeddedKafkaBroker.addTopics(missingTopics.toArray(String[]::new));
             }
-            var recordsToDelete = recordsToDelete(embeddedKafkaBroker, topics);
+            var recordsToDelete = buildRecordsToDelete(embeddedKafkaBroker, topics);
             if (!recordsToDelete.isEmpty()) {
                 adminClient.deleteRecords(recordsToDelete).all().get(WAIT_TIMEOUT_SECONDS, TimeUnit.SECONDS);
             }
         }
     }
 
-    private Map<TopicPartition, RecordsToDelete> recordsToDelete(
+    private Map<TopicPartition, RecordsToDelete> buildRecordsToDelete(
             EmbeddedKafkaBroker embeddedKafkaBroker,
             List<String> topics) {
-        try (var consumer = new KafkaConsumer<byte[], byte[]>(consumerProperties(embeddedKafkaBroker))) {
+        try (var consumer = new KafkaConsumer<byte[], byte[]>(buildConsumerProperties(embeddedKafkaBroker))) {
             var topicPartitions = topics.stream()
-                    .flatMap(topic -> topicPartitions(topic, embeddedKafkaBroker.getPartitionsPerTopic()).stream())
+                    .flatMap(topic -> buildTopicPartitions(topic, embeddedKafkaBroker.getPartitionsPerTopic()).stream())
                     .toList();
             consumer.assign(topicPartitions);
             var endOffsets = consumer.endOffsets(topicPartitions);
@@ -73,19 +73,19 @@ public class EmbeddedKafkaTestExtension implements BeforeEachCallback {
         }
     }
 
-    private List<TopicPartition> topicPartitions(String topic, int partitions) {
+    private List<TopicPartition> buildTopicPartitions(String topic, int partitions) {
         return java.util.stream.IntStream.range(0, partitions)
                 .mapToObj(partition -> new TopicPartition(topic, partition))
                 .toList();
     }
 
-    private Properties adminProperties(EmbeddedKafkaBroker embeddedKafkaBroker) {
+    private Properties buildAdminProperties(EmbeddedKafkaBroker embeddedKafkaBroker) {
         var properties = new Properties();
         properties.put(AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG, embeddedKafkaBroker.getBrokersAsString());
         return properties;
     }
 
-    private Properties consumerProperties(EmbeddedKafkaBroker embeddedKafkaBroker) {
+    private Properties buildConsumerProperties(EmbeddedKafkaBroker embeddedKafkaBroker) {
         var properties = new Properties();
         properties.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, embeddedKafkaBroker.getBrokersAsString());
         properties.put(ConsumerConfig.GROUP_ID_CONFIG, "embedded-kafka-cleanup-" + UUID.randomUUID());
