+++ src/main/java/dev/vality/testcontainers/annotations/postgresql/EmbeddedPostgresqlTestExtension.java
@@ -7,11 +7,6 @@ import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.platform.commons.support.AnnotationSupport;
-import org.springframework.boot.test.util.TestPropertyValues;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.test.context.ContextConfigurationAttributes;
-import org.springframework.test.context.ContextCustomizer;
-import org.springframework.test.context.ContextCustomizerFactory;
 
 import java.util.List;
 import java.util.Optional;
@@ -54,11 +49,11 @@ public class EmbeddedPostgresqlTestExtension implements BeforeAllCallback, Befor
         return AnnotationSupport.findAnnotation(context.getTestClass(), EmbeddedPostgresqlTest.class);
     }
 
-    private static Optional<EmbeddedPostgresqlTest> findAnnotation(Class<?> testClass) {
+    static Optional<EmbeddedPostgresqlTest> findAnnotation(Class<?> testClass) {
         return AnnotationSupport.findAnnotation(testClass, EmbeddedPostgresqlTest.class);
     }
 
-    private static EmbeddedPostgresql getOrStart(EmbeddedPostgresqlTest annotation) {
+    static EmbeddedPostgresql getOrStart(EmbeddedPostgresqlTest annotation) {
         var postgresql = THREAD_POSTGRESQL.get();
         if (postgresql == null) {
             postgresql = EmbeddedPostgresql.start(annotation);
@@ -67,43 +62,7 @@ public class EmbeddedPostgresqlTestExtension implements BeforeAllCallback, Befor
         return postgresql;
     }
 
-    public static class EmbeddedPostgresqlTestContextCustomizerFactory implements ContextCustomizerFactory {
-
-        @Override
-        public ContextCustomizer createContextCustomizer(
-                Class<?> testClass,
-                List<ContextConfigurationAttributes> configAttributes) {
-            return (context, mergedConfig) ->
-                    findAnnotation(testClass).ifPresent(annotation -> init(context, annotation));
-        }
-
-        private void init(ConfigurableApplicationContext context, EmbeddedPostgresqlTest annotation) {
-            var postgresql = getOrStart(annotation);
-            THREAD_POSTGRESQL.set(postgresql);
-            var jdbcUrl = postgresql.jdbcUrl();
-            var username = annotation.username();
-            var password = annotation.password();
-            TestPropertyValues.of(
-                            "spring.datasource.url=" + jdbcUrl,
-                            "spring.datasource.username=" + username,
-                            "spring.datasource.password=" + password,
-                            "spring.flyway.url=" + jdbcUrl,
-                            "spring.flyway.user=" + username,
-                            "spring.flyway.password=" + password,
-                            "postgres.db.url=" + jdbcUrl,
-                            "postgres.db.user=" + username,
-                            "postgres.db.username=" + username,
-                            "postgres.db.password=" + password,
-                            "flyway.url=" + jdbcUrl,
-                            "flyway.user=" + username,
-                            "flyway.password=" + password,
-                            "flyway.postgresql.transactional.lock=false")
-                    .and(annotation.properties())
-                    .applyTo(context);
-        }
-    }
-
-    private record EmbeddedPostgresql(EmbeddedPostgres delegate, String jdbcUrl) {
+    record EmbeddedPostgresql(EmbeddedPostgres delegate, String jdbcUrl) {
 
         @SneakyThrows
         private static EmbeddedPostgresql start(EmbeddedPostgresqlTest annotation) {
