+++ src/main/resources/META-INF/spring.factories
@@ -6,4 +6,4 @@ dev.vality.testcontainers.annotations.kafka.EmbeddedKafkaTestContextCustomizerFa
 dev.vality.testcontainers.annotations.clickhouse.ClickhouseTestcontainerExtension.ClickhouseTestcontainerContextCustomizerFactory,\
 dev.vality.testcontainers.annotations.minio.MinioTestcontainerExtension.MinioTestcontainerContextCustomizerFactory,\
 dev.vality.testcontainers.annotations.opensearch.OpensearchTestcontainerExtension.OpensearchTestcontainerContextCustomizerFactory,\
-dev.vality.testcontainers.annotations.postgresql.EmbeddedPostgresqlTestExtension.EmbeddedPostgresqlTestContextCustomizerFactory
+dev.vality.testcontainers.annotations.postgresql.EmbeddedPostgresqlTestContextCustomizerFactory
